Fully working anime cog
The cog fully works. However it will need refactored as the same embed is made of next airing and info command and could be put in the same place using the same function. Same with mal search and the reference. I will keep the old cog in for the time being